backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymupdf==1.24.10
python-docx==1.0.0
requests==2.32.3
rank-bm25==0.2.2